320 - Move constants and types
