+++ components/game/game.tsx
@@ -7,8 +7,8 @@ import { Pcd } from '~/lib/domain/pcd';
 import { GameTimer } from '~/components/game/game-timer';
 import { GameProgress } from '~/components/game/game-progress';
 import { ScenarioMap } from '~/components/scenario/scenario-map';
-import { ScenarioUserGame } from '~/components/usergame/usergame';
 import React from 'react';
+import { ScenarioUserGame } from '~/lib/types';
 
 type GameProps = {
     scenario: ScenarioUserGame;
